перенос тестовых массивов в utils
