+++ src/components/sorting-page/sorting-page-utils.ts
@@ -1,3 +1,5 @@
+import { ElementStates } from "../../types/element-states";
+
 export const randomSortArr = () => {
   const arrLength = Math.floor(Math.random() * 14) + 3;
   let sortArr = [];
@@ -7,3 +9,16 @@ export const randomSortArr = () => {
   }
   return sortArr;
 };
+
+export const oneElementForSort = [{ number: 18, state: ElementStates.Default }];
+
+export const elementsForSort = [
+  { number: 28, state: ElementStates.Default },
+  { number: 100, state: ElementStates.Default },
+  { number: 10, state: ElementStates.Default }
+];
+export const sorted = [
+    { number: 10, state: ElementStates.Default },
+    { number: 28, state: ElementStates.Default },
+    { number: 100, state: ElementStates.Default }
+  ];
